+++ CHANGELOG.md
@@ -6,5 +6,6 @@
 * [aesthetic-adapter-fela](./packages/aesthetic-adapter-fela/CHANGELOG.md)
 * [aesthetic-adapter-glamor](./packages/aesthetic-adapter-glamor/CHANGELOG.md)
 * [aesthetic-adapter-jss](./packages/aesthetic-adapter-jss/CHANGELOG.md)
+* [aesthetic-adapter-typestyle](./packages/aesthetic-adapter-typestyle/CHANGELOG.md)
 * [aesthetic-native](./packages/aesthetic-native/CHANGELOG.md)
 * [aesthetic-utils](./packages/aesthetic-utils/CHANGELOG.md)